refactor(step_definitions): refactor to multiple step modules
