+++ features/step_definitions/object_steps.js
@@ -3,22 +3,6 @@ var def = require('./def')
 
 module.exports = def(function(World, Given, When, Then) {
 
-  Given(/^a BMS file as follows$/, function (string) {
-    this.parseBMS(string)
-  })
-
-  Then(/^there should be (\d+) header sentences?$/, function (n) {
-    expect(this.result.headerSentences).to.equal(+n)
-  })
-
-  Then(/^there should be (\d+) channel sentences?$/, function (n) {
-    expect(this.result.channelSentences).to.equal(+n)
-  })
-
-  Then(/^the header "([^"]*)" should have value "([^"]*)"$/, function (name, value) {
-    expect(this.chart.headers.get(name)).to.equal(value)
-  })
-
   Then(/^there should be (\d+) objects$/, function (n) {
     expect(this.chart.objects.all()).to.have.length(+n)
   })
@@ -41,16 +25,6 @@ module.exports = def(function(World, Given, When, Then) {
     var objectBeat    = this.chart.measureToBeat(object.measure, object.fraction)
     var objectSeconds = this.timing.beatToSeconds(objectBeat)
     expect(objectSeconds).to.be.closeTo(+seconds, 1e-3)
-  })
-
-  Then(/^there should be (\d+) playable notes?$/, function (n, seconds) {
-    expect(this.notes.count()).to.equal(+n)
-  })
-
-  Then(/^object (\S+) should be a long note from beat ([\d\.]+) to ([\d\.]+)$/, function (value, a, b, seconds) {
-    var note = this.getNote(value)
-    expect(note.beat).to.equal(+a)
-    expect(note.endBeat).to.equal(+b)
   })
 
 })
